Update PAH_data.py
Edited  draw_molecule: added a more comprehensive description, cleaned up the if-elif-else statements, and made the function work with numpy.ndarray dtype. Ensured the list passed into MolsToGridImage is a list of strs, using the overall fastest (runtime) method I could think of. Changed the last error statement. Made the function work if legend is str while mol is an iterable by repeating the given string as legend of each molecule. Added a new argument so that the size of the image can be changed.